feat: add tested local Gen 3 candidate

+++ audio.cpp
@@ -29,7 +29,7 @@ static uint32_t gLastChirpAt = 0;
 enum AudioEventKind : uint8_t { AUDIO_EVENT_SFX = 0, AUDIO_EVENT_CHIRP };
 struct AudioEvent {
   uint8_t kind;
-  uint8_t value;
+  uint16_t value;
 };
 
 // ---- I2C del códec ----
@@ -271,7 +271,7 @@ static void playTone(const Note &note) {
   }
 }
 
-static void playSpeciesChirp(uint8_t dex) {
+static void playSpeciesChirp(uint16_t dex) {
   SpeciesChirpProfile profile{};
   if (!speciesChirpProfile(dex, &profile)) return;
   for (uint8_t i = 0; i < profile.count; i++) {
@@ -381,6 +381,6 @@ void speciesChirpPlay(int16_t dex, bool force) {
   uint32_t cooldown = force ? 220UL : 800UL;
   if (gLastChirpAt && now - gLastChirpAt < cooldown) return;
   gLastChirpAt = now;
-  AudioEvent event = { AUDIO_EVENT_CHIRP, (uint8_t)dex };
+  AudioEvent event = { AUDIO_EVENT_CHIRP, (uint16_t)dex };
   xQueueSend(gQ, &event, gMode == SOUND_FULL ? pdMS_TO_TICKS(28) : 0);
 }
